nextme/pyproject.toml
[project]
name = "nextme"
version = "0.1.0"
description = "Personal AI Life Management System"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-ai>=0.1",
    "llama-index>=0.11",
    "llama-index-vector-stores-chroma>=0.3",
    "chromadb>=0.5",
    "mem0ai>=0.1",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "typer>=0.12",
    "rich>=13.0",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "mcp>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.11",
    "ruff>=0.6",
    "types-pyyaml",
]

[project.scripts]
nextme = "src.cli.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: integration tests with real local components",
    "external: tests requiring external API credentials",
    "llm: tests making real LLM API calls",
    "slow: slow-running tests",
]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
